+++ ide-gen-proto.sh
@@ -76,4 +76,4 @@ gen_proto_js paymentservice
 gen_proto_go productcatalogservice
 # gen_proto_php quote
 gen_proto_python recommendationservice
-gen_proto_rust shippingservice
+gen_proto_rust shipping